refactor: rename module dir to list
