+++ src/pages/List/index.tsx
@@ -3,8 +3,8 @@ import { useIsFetching } from 'react-query'
 
 import { Card } from '../../components'
 import { USE_LISTS_KEY } from '../../infra'
-import { Add } from '../../modules/gifts/Add'
-import { ListView } from '../../modules/gifts/List'
+import { Add } from '../../modules/lists/Add'
+import { ListView } from '../../modules/lists/List'
 import { Layout } from '../common/layouts'
 
 const List = () => {
